backend/requirements.txt
FastAPI
uvicorn
pydantic
pydantic-settings
motor
pymongo
python-dotenv
sqlalchemy
databases
asyncpg
# Testing dependencies
pytest
pytest-asyncio
httpx